+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enhance `fullscreen` option and `play` method to support [`FullscreenOptions`](https://developer.mozilla.org/en-US/docs/Web/API/FullscreenOptions) (#482, #483).
 - Refactor the `toggle` method for toggling image between current/natural ratio (#477).
+- Improve the `toggle` method for zooming by dblclick (#422).
 
 ## 1.9.2 (May 29, 2021)